refactor: move `linkForUuid` and `linkForUuidSync` to `helper/formatting` module
