+++ module/data/item/discipline.mjs
@@ -1,8 +1,8 @@
 import ClassTemplate from "./templates/class.mjs";
 import ItemDescriptionTemplate from "./templates/item-description.mjs";
-import { linkForUuidSync } from "../../utils.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as LEGEND from "../../config/legend.mjs";
+import { linkForUuidSync } from "../../helper/formatting.mjs";
 
 const { isEmpty } = foundry.utils;
